v0.4 update!
now every button on the controls are made with icons, for pure aesthetic reasons.
the boxes are selectable with the mouse buttons
now you can delete boxes!
added the isotype to the tab icon, and the logotype to the top left corner of the app.
dark/light mode icons updated